pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "handwriting-processor"
version = "0.1.0"
authors = [
    { name="Callum Alpass", email="callumalpass@gmail.com" },
]
description = "A tool to OCR handwritten notes from PDFs."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-dotenv",
    "google-generativeai",
    "PyMuPDF",
    "Pillow",
    "Jinja2",
    "PyYAML",
    "tqdm",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "flake8",
]

[project.scripts]
handwrite = "handwriting_processor.main:main"
